Refactor collections, global data, plugins and jsx extension to their own files
