+++ config/extensions/jsx.js
@@ -1,0 +1,11 @@
+export default function setupJsxExtension(eleventyConfig) {
+  eleventyConfig.addExtension(["11ty.jsx", "11ty.ts", "11ty.tsx"], {
+    key: "11ty.js",
+    compile: function () {
+      return async function (data) {
+        let content = await this.defaultRenderer(data);
+        return renderToStaticMarkup(content);
+      };
+    },
+  });
+}
